+++ programs/dynamic-bonding-curve/src/constants.rs
@@ -102,6 +102,8 @@ pub mod fee {
     pub const MIN_POOL_CREATION_FEE: u64 = 1_000_000;
     // max pool creation fee: 100 SOL
     pub const MAX_POOL_CREATION_FEE: u64 = 100_000_000_000;
+
+    pub const PROTOCOL_LIQUIDITY_MIGRATION_FEE_BPS: u16 = 5; // 0.05%
 }
 
 pub mod seeds {