+++ programs/dynamic-bonding-curve/src/state/virtual_pool.rs
@@ -132,7 +132,10 @@ pub struct VirtualPool {
     pub is_withdraw_leftover: u8,
     /// is creator withdraw surplus
     pub is_creator_withdraw_surplus: u8,
-    /// migration fee withdraw status, first bit is for partner, second bit is for creator
+    /// migration fee withdraw status
+    /// 0 is for partner,
+    /// 1 is for creator,
+    /// 2 is for protocol claim migration fee
     pub migration_fee_withdraw_status: u8,
     /// pool metrics
     pub metrics: PoolMetrics,
@@ -150,8 +153,12 @@ pub struct VirtualPool {
     pub has_swap: u8,
     /// Padding for further use
     pub _padding_0: [u8; 5],
+    pub protocol_liquidity_migration_fee_bps: u16,
+    pub _padding_1: [u8; 6],
+    pub protocol_migration_base_fee_amount: u64,
+    pub protocol_migration_quote_fee_amount: u64,
     /// Padding for further use
-    pub _padding_1: [u64; 6],
+    pub _padding_2: [u64; 3],
 }
 
 const_assert_eq!(VirtualPool::INIT_SPACE, 416);
@@ -208,6 +215,7 @@ impl VirtualPool {
         pool_type: u8,
         activation_point: u64,
         base_reserve: u64,
+        protocol_liquidity_migration_fee_bps: u16,
     ) {
         self.volatility_tracker = volatility_tracker;
         self.config = config;
@@ -219,6 +227,7 @@ impl VirtualPool {
         self.pool_type = pool_type;
         self.activation_point = activation_point;
         self.base_reserve = base_reserve;
+        self.protocol_liquidity_migration_fee_bps = protocol_liquidity_migration_fee_bps;
     }
 
     pub fn get_swap_result_from_exact_output(
@@ -1071,7 +1080,7 @@ impl VirtualPool {
         self.migration_fee_withdraw_status.bitand(mask) == 0
     }
     pub fn update_withdraw_migration_fee(&mut self, mask: u8) {
-        self.migration_fee_withdraw_status = self.migration_fee_withdraw_status.bitxor(mask)
+        self.migration_fee_withdraw_status = self.migration_fee_withdraw_status.bitxor(mask);
     }
 
     pub fn get_migration_progress(&self) -> Result<MigrationProgress> {
@@ -1124,6 +1133,31 @@ impl VirtualPool {
         self.creation_fee_bits = self
             .creation_fee_bits
             .bitxor(PROTOCOL_CREATION_FEE_CLAIMED_MASK)
+    }
+
+    pub fn save_protocol_liquidity_migration_fee(&mut self, base_amount: u64, quote_amount: u64) {
+        self.protocol_migration_base_fee_amount = base_amount;
+        self.protocol_migration_quote_fee_amount = quote_amount;
+    }
+
+    pub fn claim_protocol_migration_fee(
+        &mut self,
+        max_base_amount: u64,
+        max_quote_amount: u64,
+    ) -> Result<(u64, u64)> {
+        let claim_base_amount = self.protocol_migration_base_fee_amount.min(max_base_amount);
+        let claim_quote_amount = self
+            .protocol_migration_quote_fee_amount
+            .min(max_quote_amount);
+
+        self.protocol_migration_base_fee_amount = self
+            .protocol_migration_base_fee_amount
+            .safe_sub(claim_base_amount)?;
+        self.protocol_migration_quote_fee_amount = self
+            .protocol_migration_quote_fee_amount
+            .safe_sub(claim_quote_amount)?;
+
+        Ok((claim_base_amount, claim_quote_amount))
     }
 }